add multithreaded implementation for load_json_data
